requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
appdirs==1.4.4
async-exit-stack==1.0.1
async-generator==1.10
Authlib==0.15.2
bcrypt==3.2.0
black==20.8b1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2
dnspython==2.0.0
ecdsa==0.14.1
email-validator==1.1.1
fastapi==0.61.1
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.9.0
httpcore==0.12.0
httplib2==0.18.1
httpx==0.16.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mypy-extensions==0.4.3
oauthlib==3.1.0
orjson==3.4.1
passlib==1.7.4
pathspec==0.8.0
promise==2.3
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydantic==1.7
pymongo==3.11.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2020.1
PyYAML==5.3.1
regex==2020.10.23
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
rsa==4.6
Rx==1.6.1
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==3.2.0
uritemplate==3.0.1
urllib3==1.25.11
uvicorn==0.11.8
websockets==8.1
